cryptoNet/requirements.txt
absl-py==0.6.1
astor==0.7.1
cycler==0.10.0
gast==0.2.0
grpcio==1.17.1
h5py==2.8.0
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
Markdown==3.0.1
matplotlib==3.0.2
mock==2.0.0
numpy==1.15.4
opencv-python==3.4.4.19
pandas==0.23.4
pbr==5.1.1
protobuf==3.6.1
pyparsing==2.3.0
python-dateutil==2.7.5
python-dotenv==0.10.1
pytz==2018.7
six==1.12.0
tb-nightly==1.13.0a20181126
tensorboard==1.12.1
termcolor==1.1.0
tf-estimator-nightly==1.12.0.dev20181217
tf-nightly-gpu==1.13.0.dev20181217
Werkzeug==0.14.1
